// src/services/dashboard/type.ts
export interface ICurrentContractData {
  bankMfo: string;
  contractDate: string;
  contractNumber: string;
  credit: number;
  debet: number;
  debit: number;
  eduContractSum: number;
  eduContractType: string;
  eduContractTypeId: number;
  eduCourse: string;
  eduCourseId: number;
  eduForm: string;
  eduFormId: number;
  eduOrganization: string;
  eduOrganizationAccount: string;
  eduOrganizationId: number;
  eduOrganizationInn: string;
  eduPeriod: number;
  eduSpeciality: string;
  eduSpecialityId: number;
  eduType: string;
  eduTypeId: number;
  eduYear: string;
  fullName: string;
  gpa: string | number;
  institutionType: string;
  pdfLink: string;
  pinfl: string;
}
// ...
export enum PaymentFormEnum {
  Grant = '11',
  Contract = '12',
}